Move runtime error checking code out of interfaces
