--- packages/core/src/errors.ts
@@ -1,4 +1,0 @@
-class CompilerError extends Error {
-  isCompilerError = true;
-}
-export class InvalidKeysError extends CompilerError {}
